Text Searching
Introduced a basic framework for searching on text. It's a bit funky with how it currently displays the find results, as it will read the  first N lines following each matching find result. Something a bit nicer will need to be implemented in the future as this is very confusing, but it works for now.

Currently only the find-text works.